Parse bulkmatches sequentially = 50x memory saving
The original way was to use json.load to read the bulkmatches.json file. This will however load the **full file** into memory, taking up up to serveral gigabytes.

This change will read the bulkmatches.json only one match at a time, and then yield it. The result is the same, but the memory usage is dropped to megabytes.

It will need the ijson library for more control over parsing the bulkmatches.json. I'm not sure whether this dependency should be listed in another file as well.